#2 More progress on -train-models CLI
This patch adds more features to our -train-models option of our CLI.

This patch adds an iterator over all the sets of hyperparameters we might want to evaluate while training our models.

This patch unstubs train_models() in main.py.

This patch gets rid of evaluate_models() since that happens during training. There's no point in adding this level of granularity.

This patch adds a lot of verbose printing so that we know which model we're currently training.

This patch adds tracking of loss in validation and training via CSV files.

This patch also adds the implies helper.

This patch does a lot more. Read the code for details.