#2 Add more model evaluation functionality.
This commit contains several feature additions.

We now print the number of the model we're on and also the total number of models. This'll let us know how far along we are.

We now track validation, training, and testing loss and accuracy.

This information is logged to a CSV and is visualized in a plot.

A summary is also saved in JSON format of the best results for the model.